refactor: split the weakening-direction policy out of the write guardrail
write_governance.py held both halves of the guardrail: the per-namespace
table deciding which direction of which key relaxes posture, and the single
enforcement loop that demands confirm + reason + actor for it. The table
grows with every setting that becomes live; the loop does not.

Also drops the two settings subscribers the subsystem reconciler replaced
from the anti-ghost manifest, which still required them to be constructed.
