+++ scripts/_ghost_wiring_manifest.txt
@@ -227,8 +227,6 @@ ENFORCED PostgresAlertRepository N/A -- constructed in api/lifecycle_helpers/met
 ENFORCED PersistentAlertSink N/A -- constructed in meta/chief_of_staff/monitor_builder.py::build_org_inflection_monitor alongside LoggingAlertSink whenever alert_repo_of(app_state) is wired; durably persists proactive alerts for GET /meta/alerts and alert_id chat-scope resolution.
 ENFORCED resolve_chat_answer N/A -- called by api/controllers/meta.py's chat() handler; routes /meta/chat to explain_alert/proposal-context-aware answering when alert_id/proposal_id is set, falling back to the free-form ask() path otherwise.
 ENFORCED ToolApprovalConsumer #2533 -- constructed in meta/toolsmith/factory.py::build_toolsmith when an approval store is present, threaded into ToolsmithCycleScheduler and driven each cycle via _consume_approved so an operator-approved tool_creation item rehydrates its blueprint, atomically claims the grant, and applies it live.
-ENFORCED DirectMcpActorSettingsSubscriber #2533 -- constructed in api/lifecycle_helpers/settings_dispatcher.py and appended to the dispatcher subscribers; rebuilds the fail-closed conversational actor on a direct_mcp_enabled change (re-checking security governance every time) so the toggle is hot-reloadable without weakening the boundary.
-ENFORCED OperatorConsoleSettingsSubscriber #2600 -- constructed in api/lifecycle_helpers/settings_dispatcher.py and appended to the dispatcher subscribers; rebuilds the fail-closed operator console on an operator_console_* change (re-checking security governance + a bound model every time) so the console toggle and its bounds are hot-reloadable without weakening the boundary.
 ENFORCED ToolInvocationTracker #2533 -- defaulted in api/construction_phase.py (ConstructionDeps.tool_invocation_tracker) so ToolsStateSlice.invocation_tracker is non-None in production boot, meeting the training-service auto-wire condition in lifecycle_runner_startup.
 ENFORCED build_web_search_provider_or_none #2595 -- resolved once in workers/runtime_builder.build_runtime_services and threaded into _build_tool_registry + _build_runtime_coordinator, and independently called from api/lifecycle_helpers/feature_wiring for the research web source, so the native web-search provider is constructed at boot.
 ENFORCED PlanningToolProvider #2595 -- constructed in workers/_coordinator_assembly._build_runtime_coordinator when a web-search provider or an agent-memory backend is present and passed as decomposition_tool_provider into build_coordinator, which threads it into AgentSessionDecompositionStrategy so the owner-run planning session gets a live web_search and/or read-only search_memory grant.
